move 'isExecutableFile' func to filetype.go
