+++ filetype_windows_test.go
@@ -1,0 +1,58 @@
+// +build windows
+
+package main
+
+func TestIsExecutableFile(t *testing.T) {
+	tests := []struct {
+		desc string
+		path string
+		want bool
+	}{
+		{
+			desc: "ok: executable batch file",
+			path: filepath.Join(testDir, "script.bat"),
+			want: true,
+		},
+		{
+			desc: "ok: executable binary (linux)",
+			path: filepath.Join(testDir, "bin"),
+			want: true,
+		},
+		{
+			desc: "ok: executable binary (windows)",
+			path: filepath.Join(testDir, "bin.exe"),
+			want: true,
+		},
+		// TODO:
+		// {
+		// 	desc: "ok: executable binary (darwin)",
+		// 	path: filepath.Join(testDir, "darwin"),
+		// 	want: true,
+		// },
+		{
+			desc: "ng: text file",
+			path: filepath.Join(testDir, "text.txt"),
+			want: false,
+		},
+		{
+			desc: "ng: directory",
+			path: testDir,
+			want: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.desc, func(t *testing.T) {
+			assert := assert.New(t)
+
+			file, err := os.Open(tt.path)
+			assert.NoError(err)
+			defer file.Close()
+			fi, err := file.Stat()
+			assert.NoError(err)
+
+			got, err := IsExecutableFile(fi, tt.path)
+			assert.Equal(tt.want, got)
+			assert.NoError(err)
+		})
+	}
+}
